+++ HISTORY.md
@@ -4,6 +4,7 @@
   * Fixed "Allow launching" when the file doesn't have executable flag (Sergio Costas)
   * Ignore SPACE key to start a search (Laurentiu-Andrei Postole)
   * Use CSS to make the window transparent (Sundeep Mediratta)
+  * Removed "ask to execute" dialog; now shows "Run as a program" (Sergio Costas)
 
 * Version 38
   * Fixed Paste in Gnome Shell 3.38 (Sergio Costas)